move tests over to crate-def-map
