+++ crates/ra_hir/src/marks.rs
@@ -1,6 +1,6 @@
 test_utils::marks!(
     name_res_works_for_broken_modules
-    item_map_enum_importing
+    can_import_enum_variant
     type_var_cycles_resolve_completely
     type_var_cycles_resolve_as_possible
     type_var_resolves_to_int_var
