+++ crates/ra_hir/src/nameres/crate_def_map.rs
@@ -273,7 +273,7 @@ impl CrateDefMap {
                 }
                 ModuleDef::Enum(e) => {
                     // enum variant
-                    tested_by!(item_map_enum_importing);
+                    tested_by!(can_import_enum_variant);
                     match e.variant(db, &segment.name) {
                         Some(variant) => PerNs::both(variant.into(), variant.into()),
                         None => {
