+++ crates/ra_hir/src/nameres/crate_def_map/tests.rs
@@ -1,3 +1,7 @@
+mod macros;
+mod globs;
+mod incremental;
+
 use std::sync::Arc;
 
 use ra_db::SourceDatabase;
@@ -62,6 +66,8 @@ fn crate_def_map_smoke_test() {
         //- /lib.rs
         mod foo;
         struct S;
+        use crate::foo::bar::E;
+        use self::E::V;
 
         //- /foo/mod.rs
         pub mod bar;
@@ -74,9 +80,13 @@ fn crate_def_map_smoke_test() {
     );
     assert_snapshot_matches!(map, @r###"
 crate
+V: t v
+E: t
+foo: t
 S: t v
 
 crate::foo
+bar: t
 f: v
 
 crate::foo::bar
@@ -87,91 +97,96 @@ E: t
 }
 
 #[test]
-fn macro_rules_are_globally_visible() {
+fn use_as() {
     let map = def_map(
         "
         //- /lib.rs
-        macro_rules! structs {
-            ($($i:ident),*) => {
-                $(struct $i { field: u32 } )*
-            }
-        }
-        structs!(Foo);
-        mod nested;
+        mod foo;
 
-        //- /nested.rs
-        structs!(Bar, Baz);
+        use crate::foo::Baz as Foo;
+
+        //- /foo/mod.rs
+        pub struct Baz;
         ",
     );
-    assert_snapshot_matches!(map, @r###"
+    assert_snapshot_matches!(map,
+        @r###"
 crate
 Foo: t v
+foo: t
 
-crate::nested
-Bar: t v
+crate::foo
 Baz: t v
-"###);
+"###
+    );
 }
 
 #[test]
-fn macro_rules_can_define_modules() {
+fn use_trees() {
     let map = def_map(
         "
         //- /lib.rs
-        macro_rules! m {
-            ($name:ident) => { mod $name;  }
-        }
-        m!(n1);
+        mod foo;
 
-        //- /n1.rs
-        m!(n2)
-        //- /n1/n2.rs
-        struct X;
+        use crate::foo::bar::{Baz, Quux};
+
+        //- /foo/mod.rs
+        pub mod bar;
+
+        //- /foo/bar.rs
+        pub struct Baz;
+        pub enum Quux {};
         ",
     );
-    assert_snapshot_matches!(map, @r###"
+    assert_snapshot_matches!(map,
+        @r###"
 crate
+Quux: t
+Baz: t v
+foo: t
 
-crate::n1
+crate::foo
+bar: t
 
-crate::n1::n2
-X: t v
-"###);
+crate::foo::bar
+Quux: t
+Baz: t v
+"###
+    );
 }
 
 #[test]
-fn macro_rules_from_other_crates_are_visible() {
-    let map = def_map_with_crate_graph(
+fn re_exports() {
+    let map = def_map(
         "
-        //- /main.rs
-        foo::structs!(Foo, Bar)
-        mod bar;
-
-        //- /bar.rs
-        use crate::*;
-
         //- /lib.rs
-        #[macro_export]
-        macro_rules! structs {
-            ($($i:ident),*) => {
-                $(struct $i { field: u32 } )*
-            }
-        }
-        ",
-        crate_graph! {
-            "main": ("/main.rs", ["foo"]),
-            "foo": ("/lib.rs", []),
-        },
-    );
-    assert_snapshot_matches!(map, @r###"
-crate
-Foo: t v
-Bar: t v
+        mod foo;
 
-crate::bar
-Foo: t v
-Bar: t v
-"###);
+        use self::foo::Baz;
+
+        //- /foo/mod.rs
+        pub mod bar;
+
+        pub use self::bar::Baz;
+
+        //- /foo/bar.rs
+        pub struct Baz;
+        ",
+    );
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Baz: t v
+foo: t
+
+crate::foo
+bar: t
+Baz: t v
+
+crate::foo::bar
+Baz: t v
+"###
+    );
 }
 
 #[test]
@@ -203,31 +218,8 @@ Baz: t v
 }
 
 #[test]
-fn glob_across_crates() {
-    covers!(glob_across_crates);
-    let map = def_map_with_crate_graph(
-        "
-        //- /main.rs
-        use test_crate::*;
-
-        //- /lib.rs
-        pub struct Baz;
-        ",
-        crate_graph! {
-            "main": ("/main.rs", ["test_crate"]),
-            "test_crate": ("/lib.rs", []),
-        },
-    );
-    assert_snapshot_matches!(map, @r###"
-crate
-Baz: t v
-"###
-    );
-}
-
-#[test]
-fn item_map_enum_importing() {
-    covers!(item_map_enum_importing);
+fn can_import_enum_variant() {
+    covers!(can_import_enum_variant);
     let map = def_map(
         "
         //- /lib.rs
@@ -244,22 +236,293 @@ E: t
 }
 
 #[test]
-fn glob_enum() {
-    covers!(glob_enum);
+fn edition_2015_imports() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /main.rs
+        mod foo;
+        mod bar;
+
+        //- /bar.rs
+        struct Bar;
+
+        //- /foo.rs
+        use bar::Bar;
+        use other_crate::FromLib;
+
+        //- /lib.rs
+        struct FromLib;
+        ",
+        crate_graph! {
+            "main": ("/main.rs", "2015", ["other_crate"]),
+            "other_crate": ("/lib.rs", "2018", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+bar: t
+foo: t
+
+crate::bar
+Bar: t v
+
+crate::foo
+FromLib: t v
+Bar: t v
+"###
+    );
+}
+
+#[test]
+fn module_resolution_works_for_non_standard_filenames() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /my_library.rs
+        mod foo;
+        use self::foo::Bar;
+
+        //- /foo/mod.rs
+        pub struct Bar;
+        ",
+        crate_graph! {
+            "my_library": ("/my_library.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Bar: t v
+foo: t
+
+crate::foo
+Bar: t v
+"###
+    );
+}
+
+#[test]
+fn name_res_works_for_broken_modules() {
+    covers!(name_res_works_for_broken_modules);
     let map = def_map(
         "
         //- /lib.rs
-        enum Foo {
-            Bar, Baz
-        }
-        use self::Foo::*;
+        mod foo // no `;`, no body
+
+        use self::foo::Baz;
+
+        //- /foo/mod.rs
+        pub mod bar;
+
+        pub use self::bar::Baz;
+
+        //- /foo/bar.rs
+        pub struct Baz;
         ",
     );
-    assert_snapshot_matches!(map, @r###"
+    assert_snapshot_matches!(map,
+        @r###"
 crate
-Foo: t
-Bar: t v
+Baz: _
+"###
+    );
+}
+
+#[test]
+fn item_map_using_self() {
+    let map = def_map(
+        "
+        //- /lib.rs
+        mod foo;
+        use crate::foo::bar::Baz::{self};
+        //- /foo/mod.rs
+        pub mod bar;
+        //- /foo/bar.rs
+        pub struct Baz;
+        ",
+    );
+    assert_snapshot_matches!(map,
+        @r###"
+crate
 Baz: t v
+foo: t
+
+crate::foo
+bar: t
+
+crate::foo::bar
+Baz: t v
+"###
+    );
+}
+
+#[test]
+fn item_map_across_crates() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /main.rs
+        use test_crate::Baz;
+
+        //- /lib.rs
+        pub struct Baz;
+        ",
+        crate_graph! {
+            "main": ("/main.rs", ["test_crate"]),
+            "test_crate": ("/lib.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Baz: t v
+"###
+    );
+}
+
+#[test]
+fn extern_crate_rename() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /main.rs
+        extern crate alloc as alloc_crate;
+
+        mod alloc;
+        mod sync;
+
+        //- /sync.rs
+        use alloc_crate::Arc;
+
+        //- /lib.rs
+        struct Arc;
+        ",
+        crate_graph! {
+            "main": ("/main.rs", ["alloc"]),
+            "alloc": ("/lib.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Arc: t v
+"###
+    );
+}
+
+#[test]
+fn extern_crate_rename_2015_edition() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /main.rs
+        extern crate alloc as alloc_crate;
+
+        mod alloc;
+        mod sync;
+
+        //- /sync.rs
+        use alloc_crate::Arc;
+
+        //- /lib.rs
+        struct Arc;
+        ",
+        crate_graph! {
+            "main": ("/main.rs", "2015", ["alloc"]),
+            "alloc": ("/lib.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Arc: t v
+"###
+    );
+}
+
+#[test]
+fn import_across_source_roots() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /lib.rs
+        pub mod a {
+            pub mod b {
+                pub struct C;
+            }
+        }
+
+        //- root /main/
+
+        //- /main/main.rs
+        use test_crate::a::b::C;
+        ",
+        crate_graph! {
+            "main": ("/main/main.rs", ["test_crate"]),
+            "test_crate": ("/lib.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+C: t v
+"###
+    );
+}
+
+#[test]
+fn reexport_across_crates() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /main.rs
+        use test_crate::Baz;
+
+        //- /lib.rs
+        pub use foo::Baz;
+
+        mod foo;
+
+        //- /foo.rs
+        pub struct Baz;
+        ",
+        crate_graph! {
+            "main": ("/main.rs", ["test_crate"]),
+            "test_crate": ("/lib.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Baz: t v
+"###
+    );
+}
+
+#[test]
+fn values_dont_shadow_extern_crates() {
+    let map = def_map_with_crate_graph(
+        "
+        //- /main.rs
+        fn foo() {}
+        use foo::Bar;
+
+        //- /foo/lib.rs
+        pub struct Bar;
+        ",
+        crate_graph! {
+            "main": ("/main.rs", ["foo"]),
+            "foo": ("/foo/lib.rs", []),
+        },
+    );
+
+    assert_snapshot_matches!(map,
+        @r###"
+crate
+Bar: t v
+foo: v
 "###
     );
 }
